refactor: move CustomRepositoryPersistenceBean to extensions.sql.repository.spi and update package imports
